add options to gcal_api: ignore_attrs_num, size_diff_threshold
these options are not exposed in the webapp yet

ignore_attrs_num: gcal_api has test to ensure it does not move wrong event:
if the event-to-keep has less attributes/fields than any event-to-move,
previously that group would be skipped unconditionally.  now, this is
configurable (ignore number of attributes or not).

size_diff_threshold: this is another test like above.  certain attributes
of each event are measured for overall text length.  not all attributes
because some of them vary randomly, or for no significant reason at
least.  (e.g., event id is sometimes 3-4 times longer, often for what
seems to be original event.  in the future maybe we should do a test
looking just at id length to determine original event.)
size_diff_threshold is the threshold under which events were previously
not moved if they failed this test.  previously it was hard-coded to 85%
(if event-to-keep was less than 85% as "big" (again, just text length)
it failed the test.  now this threshold is configurable, or it can be
set to 0 to allow any size difference.